fix default names config bug, and moved logic to its own folder
