+++ animated-transformer/src/app/activation-vis/activation-vis.component.spec.ts
@@ -13,13 +13,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ActivationVisComponent } from './activation-vis.component';
 
 import { CommonModule } from '@angular/common';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
@@ -34,10 +33,11 @@ import { CodemirrorConfigEditorModule } from '../codemirror-config-editor/codemi
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { AxisWrapperComponent } from './axis-wrapper/axis-wrapper.component';
-import { NanValidatorDirective } from '../form-validators/nan-validator.directive'
+import { NanValidatorDirective } from '../form-validators/nan-validator.directive';
 import { BoundedFloatValidatorDirective } from '../form-validators/bounded-float-validator.directive';
 import { ActivationManagerDirective } from './activation-manager.directive';
 import { CornerActivationComponent } from './corner-activation/corner-activation.component';
+import { AutoCompletedTextInputComponent } from '../auto-completed-text-input/auto-completed-text-input.component';
 
 describe('ActivationVisComponent', () => {
   let component: ActivationVisComponent;
@@ -46,8 +46,8 @@ describe('ActivationVisComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [
+        NoopAnimationsModule,
         CommonModule,
-        BrowserAnimationsModule,
         FormsModule,
         ReactiveFormsModule,
         RouterTestingModule,
@@ -63,6 +63,7 @@ describe('ActivationVisComponent', () => {
         // ---
         CodemirrorConfigEditorModule,
         TensorImageModule,
+        AutoCompletedTextInputComponent,
       ],
       declarations: [
         ActivationVisComponent,
@@ -70,8 +71,8 @@ describe('ActivationVisComponent', () => {
         ActivationManagerDirective,
         CornerActivationComponent,
         NanValidatorDirective,
-        BoundedFloatValidatorDirective
-      ]
+        BoundedFloatValidatorDirective,
+      ],
     }).compileComponents();
   });
